rename test/coverage_specs2 -> test/coverage_specs2_with_junit
