+++ test/coverage_specs2_with_junit/B1.scala
@@ -1,4 +1,4 @@
-package coverage_specs2
+package coverage_specs2_with_junit
 
 object B1 {
