+++ test/coverage_specs2_with_junit/B2.java
@@ -1,5 +1,5 @@
-package coverage_specs2;
-
+package coverage_specs2_with_junit;
+import coverage_specs2_with_junit.C2;
 class B2 {
     public static String b2_a() {
         return C2.c2("hello from b2_a");
