+++ test/coverage_specs2_with_junit/BUILD
