+++ test/coverage_specs2_with_junit/C2.scala
@@ -1,4 +1,4 @@
-package coverage_specs2
+package coverage_specs2_with_junit
 
 object C2 {
   def c2(input: String): String =
