+++ test/coverage_specs2_with_junit/D1.scala
@@ -1,4 +1,4 @@
-package coverage_specs2
+package coverage_specs2_with_junit
 
 object D1 {
   def veryLongFunctionNameIsHereAaaaaaaaa(): Unit = {
