+++ test_coverage.sh
@@ -9,5 +9,5 @@ fi
 
 test_dir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )/test/shell
 
-. "${test_dir}"/test_coverage_specs2.sh
+. "${test_dir}"/test_coverage_specs2_with_junit.sh
 . "${test_dir}"/test_coverage_scalatest.sh
